Another piece of the reconciliation puzzle (#257)
This covers the next piece of reconciliation where we actually copy
extent files from one downstairs to another.

Upstairs:
Update the repair logic to now do more than just close and then
re-open extents. The upstairs will now build the complete repair
steps needed to repair an extent.  This includes the logic about
which downstairs should receive which repair commands.
Added big picture comment on repair process.

Protocol:
New ExtentFlush to flush a specific extent.
New ExtentRepair command for repairing a specific extent.
Replaced a few specific ACKs with a generic repair ACK.

Downstairs:
Made repair mode always on, and starting at port + REPAIR_PORT_OFFSET.
New repair-api command that can dump a json file for the repair-client API
Added the logic to support the new repair messages.

Downstairs region:
The downstairs coordination of a repair live in this file
New functions to get the names for the new repair directories and
a bunch of tests around them.  We can now get names for specific
parts of the extent, the housing directory, the copy, replace, and
completed directories, as well as provide an optional file extension
to the file name.
During repair, we check for and remove any copy or completed directories.
If there is a replay directory, then the extent open will do the
replay and replace existing extent files, only if the open is read/write.
New command to just flush a specific extent.

Downstairs repair:
This new file is the dropshot server for extent repair requests.
It listens on the specified port and serves (anyone currently) the
list of files for an extent, or one of the actual extent files.

Repair-client:
Generate a library using progenitor that is used by a downstairs to
request repair extents from another downstairs. This library generation
uses the json file that is output from the downstairs repair-api command.

Tests:
More tests for repair
More tests for building repair work list.
A New test that will in a loop:
 Create a workload that will produce downstairs out of sync
 Restart and repair the out of sync downstairs.
Improved usage messages for create-generic-ds.sh

+++ upstairs/src/mend.rs
@@ -34,7 +34,7 @@ pub struct ExtentFix {
 #[derive(Debug)]
 pub struct DownstairsMend {
     // Index by extent ID
-    pub mend: HashMap<u64, ExtentFix>,
+    pub mend: HashMap<usize, ExtentFix>,
 }
 
 impl DownstairsMend {
@@ -108,7 +108,7 @@ impl DownstairsMend {
             if *dirty0 || c1.dirty[i] || c2.dirty[i] {
                 println!("Extents {} dirty", i);
                 let ef = make_repair_list(i, c0, c1, c2);
-                dsm.mend.insert(i as u64, ef);
+                dsm.mend.insert(i, ef);
             } else {
                 to_check.push(i as usize);
             }
@@ -126,7 +126,7 @@ impl DownstairsMend {
             {
                 println!("Extent {} has flush number mismatch", i);
                 let ef = make_repair_list(*i, c0, c1, c2);
-                dsm.mend.insert(*i as u64, ef);
+                dsm.mend.insert(*i, ef);
             } else {
                 second_check.push(*i);
             }
@@ -142,7 +142,7 @@ impl DownstairsMend {
             {
                 println!("generation number mismatch {}", i);
                 let ef = make_repair_list(*i, c0, c1, c2);
-                dsm.mend.insert(*i as u64, ef);
+                dsm.mend.insert(*i, ef);
             }
         }
